feat: GlobalExceptionHandler added

+++ src/main/java/com/hotel/roomBooker/service/HotelService.java
@@ -4,6 +4,8 @@ import com.hotel.roomBooker.DTO.HotelMapper;
 import com.hotel.roomBooker.DTO.HotelRequestDTO;
 import com.hotel.roomBooker.DTO.HotelResponseDTO;
 import com.hotel.roomBooker.DTO.PaginatedHotelResponseDTO;
+import com.hotel.roomBooker.exception.InvalidInputException;
+import com.hotel.roomBooker.exception.ResourceNotFoundException;
 import com.hotel.roomBooker.model.Hotel;
 import com.hotel.roomBooker.repository.HotelRepository;
 import lombok.AllArgsConstructor;
@@ -12,6 +14,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import java.util.List;
 import java.util.Optional;
+
 
 @Service
 @AllArgsConstructor
@@ -31,30 +34,49 @@ public class HotelService {
         return response;
     }
 
-    public Optional<HotelResponseDTO> getHotelById(Long id) {
-        return hotelRepository.findById(id)
-                .map(hotelMapper::toDTO);
+    public HotelResponseDTO getHotelById(Long id) {
+        Hotel hotel = hotelRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Hotel with id " + id + " not found"));
+        return hotelMapper.toDTO(hotel);
     }
 
+
     public HotelResponseDTO createHotel(HotelRequestDTO hotelRequest) {
+        Optional<Hotel> existingHotel = hotelRepository.findByNameAndCityAndAddress(
+                hotelRequest.getName(),
+                hotelRequest.getCity(),
+                hotelRequest.getAddress()
+        );
+        if (existingHotel.isPresent()) {
+            throw new InvalidInputException("Hotel with the same name, city, and address already exists");
+        }
         Hotel hotel = hotelMapper.toEntity(hotelRequest);
         Hotel createdHotel = hotelRepository.save(hotel);
         return hotelMapper.toDTO(createdHotel);
     }
 
+
     public HotelResponseDTO updateHotel(Long id, HotelRequestDTO hotelRequest) {
         Hotel hotel = hotelRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Hotel not found"));
-        hotel.setName(hotelRequest.getName());
-        hotel.setAnnouncementTitle(hotelRequest.getAnnouncementTitle());
-        hotel.setCity(hotelRequest.getCity());
-        hotel.setAddress(hotelRequest.getAddress());
-        hotel.setDistanceFromCityCenter(hotelRequest.getDistanceFromCityCenter());
-        Hotel updatedHotel = hotelRepository.save(hotel);
-        return hotelMapper.toDTO(updatedHotel);
+                .orElseThrow(() -> new ResourceNotFoundException("Hotel with id " + id + " not found"));
+
+        try {
+            hotel.setName(hotelRequest.getName());
+            hotel.setAnnouncementTitle(hotelRequest.getAnnouncementTitle());
+            hotel.setCity(hotelRequest.getCity());
+            hotel.setAddress(hotelRequest.getAddress());
+            hotel.setDistanceFromCityCenter(hotelRequest.getDistanceFromCityCenter());
+            Hotel updatedHotel = hotelRepository.save(hotel);
+            return hotelMapper.toDTO(updatedHotel);
+        } catch (Exception e) {
+            throw new InvalidInputException("Invalid data provided for hotel update");
+        }
     }
 
     public void deleteHotel(Long id) {
+        if (!hotelRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Hotel with id " + id + " not found");
+        }
         hotelRepository.deleteById(id);
     }
 }
